♻️ Move StmtOptions and DefaultStmtOptions to opts
